Move User and UserList into component files
