+++ src/main/js/app.js
@@ -6,6 +6,7 @@
 const React = require('react');
 const ReactDOM = require('react-dom');
 const client = require('./client');
+import UserList from "./components/UserList";
 
 class App extends React.Component {
 
@@ -24,39 +25,6 @@ class App extends React.Component {
     render() {
         return (
             <UserList users={this.state.users}/>
-        )
-    }
-}
-
-class UserList extends React.Component {
-    render() {
-        console.debug(this.props)
-        const users = this.props.users.map(user =>
-            <User key={user.email} user={user}/>
-        );
-        return (
-            <table>
-                <tbody>
-                <tr>
-                    <th>First Name</th>
-                    <th>Last Name</th>
-                    <th>email</th>
-                </tr>
-                {users}
-                </tbody>
-            </table>
-        )
-    }
-}
-
-class User extends React.Component {
-    render() {
-        return (
-            <tr>
-                <td>{this.props.user.firstName}</td>
-                <td>{this.props.user.lastName}</td>
-                <td>{this.props.user.email}</td>
-            </tr>
         )
     }
 }
