+++ src/main/js/components/User.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+
+class User extends React.Component {
+    render() {
+        return (
+            <tr>
+                <td>{this.props.user.firstName}</td>
+                <td>{this.props.user.lastName}</td>
+                <td>{this.props.user.email}</td>
+            </tr>
+        )
+    }
+}
+
+export default User
