+++ src/main/resources/templates/index.html
@@ -2,7 +2,7 @@
 <html xmlns:th="https://www.thymeleaf.org">
 <head lang="en">
     <meta charset="UTF-8"/>
-    <title>ReactJS + Spring Data REST</title>
+    <title>Award Portal</title>
     <link rel="stylesheet" href="/main.css" />
 </head>
 <body>
